Alert

+++ components/FavoritesComponent.js
@@ -6,14 +6,19 @@ import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
+import { FlatList, View, Text, Alert } from 'react-native';
+
+
 const mapStateToProps = state => {
     return {
         campsites: state.campsites,
         favorites: state.favorites
     };
 };
-const mapDispatchToProps = { deleteFavorite: campsiteId => (deleteFavorite(campsiteId)) };
 
+const mapDispatchToProps = {
+    deleteFavorite: campsiteId => (deleteFavorite(campsiteId))
+};
 
 class Favorites extends Component {
 
@@ -28,7 +33,24 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite campsite ' + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
 
@@ -41,6 +63,7 @@ class Favorites extends Component {
                         onPress={() => navigate('CampsiteInfo', { campsiteId: item.id })}
                     />
                 </Swipeout>
+
             );
         };
 
@@ -66,4 +89,4 @@ class Favorites extends Component {
     }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Favorites);
+export default connect(mapStateToProps, mapDispatchToProps)(Favorites);
